fix howDiscoverPicker
