+++ src/components/cartPage/howDiscoverPicker/networksList.ts
@@ -1,0 +1,29 @@
+type Network = {
+  name: string;
+  value: string;
+};
+
+const networksList: Network[] = [
+  {
+    name: 'Tiktok',
+    value: 'Tiktok',
+  },
+  {
+    name: 'Linkedin',
+    value: 'Linkedin',
+  },
+  {
+    name: 'Instagram',
+    value: 'Instagram',
+  },
+  {
+    name: 'Facebook',
+    value: 'Facebook',
+  },
+  {
+    name: 'Other',
+    value: '',
+  },
+];
+
+export default networksList;
